requirements.txt
# EfficientDet
tqdm==4.62.2
lxml==4.6.1
absl-py==0.10.0
matplotlib==3.2.2
numpy==1.19.4
Pillow==8.0.0
PyYAML==5.3.1
six==1.15.0
tensorflow==2.5.0
tensorflow-addons==0.12
tensorflow-hub==0.11
neural-structured-learning==1.3.1
tensorflow-model-optimization==0.5
Cython==0.29.13
git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI


# Yolov5
opencv-python==4.1.2.30
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tensorboard>=2.4.1
seaborn>=0.11.0
pandas
thop  # FLOPs computation
